% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALA_record_download.R
\name{ala_record_download}
\alias{ala_record_download}
\title{Download ALA records identified in count search from species_list}
\usage{
ala_record_download(species_list, common_names, path)
}
\arguments{
\item{species_list}{A `data.frame` containing columns of 'correct_name', 'search_term', 'common_name' for species, and T/F columns for each list}

\item{filter_df}{A `list` of ALA query conditions}
}
\description{
Redoing Martin's function minus lapply and to hopefully run smoother
}
