% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aus_territory.R
\docType{data}
\name{aus_territory}
\alias{aus_territory}
\title{Australian Territory .shp file}
\format{
## `aus_territory`
A data frame with 1 row and 1 column - a multipolygon object describing all
   terrestrial and maritime teritory of Australia.
}
\source{
<https://ecat.ga.gov.au/geonetwork/srv/api/records/38337c55-6fdd-43b7-a134-a2511d4a6eff>
}
\usage{
aus_territory
}
\description{
A data-frame containing the multipolygon object that describes the maritime
   and terrestrial regions of Australia as defined by the Seas and Submerged
   Lands Act (SSLA) 1973. The SSLA 1973 Act defines a number of different
   maritime boundaries, including but not limited to the Exclusive Economic
   Zone (EEZ) as per the Perth treaty, the Contiguous Zone, and the
   Territorial Seas. All territories contained within these borders fall
   under some degree of control of Australia and are therefore of ecological
   and biological importance to Australia.
The terrestrial regions included in this .shp file include mainland Australia,
   Tasmania, island territories belonging to Australian states and
   territories (inc. Lord Howe Island (NSW) and Macquarie Island (TAS)),
   Norfolk Island, Ashmore & Cartier Islands, Christmas Island, Cocos Keeling
   Island, Macquarie Island, and the Heard & McDonald Islands. The maritime
   jurisdictions include any waters up to 200NM away from any of these named
   land masses. Note that neither the Australian Antarctic Territory nor its
   associated waters have been included here.
}
\keyword{datasets}
