% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_ALA_query.R
\name{build_ala_query}
\alias{build_ala_query}
\title{Build ALA query object for searching species}
\usage{
build_ala_query(start_days_ago, lat_bounds, lng_bounds)
}
\arguments{
\item{start_days_ago}{A `dbl` objecting indicating how many days prior would we would like to search from}

\item{lat_bounds}{A 'dbl' vector indicating the lower and upper latitude limits of the query}

\item{lng_bounds}{A 'dbl' vector indicating the lower and upper longitude limits of the query}
}
\description{
Redoing Martin's function minus lapply and to hopefully run smoother
}
