% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ala_records.R
\name{build_galah_query}
\alias{build_galah_query}
\title{Set up filtering parameters for getting species counts and records with {galah}}
\usage{
build_galah_query(lat_bounds, lng_bounds, start_days_ago, end_days_ago = 0)
}
\arguments{
\item{lat_bounds}{A `numeric` vector indicating lower and upper latitude bounds}

\item{lng_bounds}{A `numeric` vector indicating lower and upper longitude bounds}

\item{start_days_ago}{A `dbl` indicating how many days prior to search from}

\item{end_days_ago}{A `dbl` indicating how many days prior to search up to.
Defaults to 0 (up to current time).}
}
\value{
A tibble containing parameters for filtering counts and records, with
   columns of `variable`, `logical`, `value`, and `query`.
}
\description{
This function is a wrapper for the {galah} function `galah::filter()`, producing
   a modifiable filter that can be used with other {galah} functions to search
   the ALA and similar biodiversity databases for species occurrences within
   some geographical bounding box and timeframe. The resulting output can be
   passed to `lookup_species_counts()` and `download_records()` to search and
   download species occurrences within the filter.
}
