% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_email.R
\name{build_gt_table}
\alias{build_gt_table}
\title{Build an HTML table containing the ALA occurrences provided in a dataframe}
\usage{
build_gt_table(df, cache_path)
}
\arguments{
\item{df}{A data.frame produced by `download_occurrences()` that is filtered
down to a given list in `build_email()`. Contains __ data columns and
logical columns for each list of interest.}

\item{cache_path}{A single string containing the path to the temporary
cache folder in which species images and maps are saved.  Must begin with
`"./"` and end  with `"/"`. When this function it is called it should contain
a `species_images` and a `maps` directory containing relevant images.}
}
\value{
A tibble that is passed on to some RMarkdown (.Rmd) file to be
   rendered as a gt table. Contains four columns: `species`, `observation`,
   `location` and `image`.
}
\description{
This implementation uses {gt} to create an HTML table of ALA occurrences from
   a provided list. This function is used within `build_email()` to produce
   interactive html tables which link to various web pages and contain
   observation images, maps and details
}
