% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_map_thumbnail.R
\name{build_map_thumbnail}
\alias{build_map_thumbnail}
\title{internal function to build maps}
\usage{
build_map_thumbnail(list_row)
}
\description{
use basemaps from OpenStreetMaps, called via `maptiles`
}
