% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koel_defensive.R
\name{check_date}
\alias{check_date}
\title{Internal checks to run on provided dates}
\usage{
check_date(var, var_name)
}
\arguments{
\item{var}{variable value to be checked}

\item{var_name}{name of variable value to be checked as a string}
}
\description{
Internal checks to run on provided dates
}
