% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koel_defensive.R
\name{check_df}
\alias{check_df}
\title{Internal checks to run on provided data.frames / tibbles}
\usage{
check_df(var, var_name, req_cols = NULL)
}
\arguments{
\item{var}{variable value to be checked}

\item{var_name}{name of variable value to be checked as a string}

\item{req_cols}{character vector of required column names for var}
}
\description{
Internal checks to run on provided data.frames / tibbles
}
