% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koel_defensive.R
\name{check_dir_path}
\alias{check_dir_path}
\title{Internal checks to run on provided directory paths}
\usage{
check_dir_path(var, var_name)
}
\arguments{
\item{var}{variable value to be checked}

\item{var_name}{name of variable value to be checked as a string}
}
\description{
Internal checks to run on provided directory paths
}
