% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koel_defensive.R
\name{check_file_path}
\alias{check_file_path}
\title{Internal checks to run on provided file paths}
\usage{
check_file_path(var, var_name, file_ext)
}
\arguments{
\item{var}{variable value to be checked}

\item{var_name}{name of variable value to be checked as a string}

\item{file_ext}{string beginning with "." specifying file extension of var}
}
\description{
Internal checks to run on provided file paths
}
