% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_species_lists2.R
\name{clean_names}
\alias{clean_names}
\title{Clean up name columns}
\usage{
clean_names(name)
}
\arguments{
\item{name}{`character string` object or column/vector to be cleaned}
}
\value{
A cleaned `character string` of the same type and length as `name`.
}
\description{
Internal function to perform numerous regex substitution that clean up
   name columns to be suitable for searching with galah
}
