% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collate_lists.R
\name{collate_lists}
\alias{collate_lists}
\title{Combine user-supplied lists into data.frame}
\usage{
collate_lists(path, list_suffix = "_list")
}
\arguments{
\item{path}{Path to the directory where lists are saved, beginning with "./"
and ending in "/".}

\item{list_suffix}{Character string between the list name and file extension.}
}
\value{
A data.frame containing four columns, with information about the
  source and location of user-supplied lists.
}
\description{
Alerts need to be run against multiple user-supplied lists. This function
  combines all existing lists and returns a single data.frame, which may
  then be passed as an argument to `get_species_lists()`.
}
