% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ala_records.R
\name{download_records}
\alias{download_records}
\title{Download ALA records for species identified with `lookup_species_count()`}
\usage{
download_records(species_list, common_names, filter_df, path)
}
\arguments{
\item{species_list}{A data.frame or tibble containing information on species
of biosecurity interest. Must contain character columns 'correct_name',
'search_term', and 'common_name' for each species, and logical columns for
each alerts list indicating whether or not the species appears on a list.
May be produced by `get_species_list2()`.}

\item{common_names}{A data.frame or tibble containing two character columns,
'correct_name' and 'common_name', and rows indicating the accepted common
name for each "correct" species name. May be produced by
`common_names_assigned()`.}

\item{filter_df}{A tibble specifying {galah} query parameters, as produced
by `galah::galah_filter()`. May be produced by `build_galah_query()`.}

\item{path}{A character string specifying the (temporary) cache directory to
save downloaded media files for each occurrence, and the downloaded occurrence
data. The string must end in "/". The path must describe an existing directory,
and if no 'species_images' folder exists within this directory then one will
be created, in which the media output will be saved.}
}
\value{
A tibble containing the downloaded data for each occurrence record.
   Contains 30 ALA-specific columns with data regarding location, media,
   uploading user and data type, 2 columns containing correct and common names,
   and logical columns for each list indicating the presence of a species
   on that list.
}
\description{
This function downloads data, including media, for species identified by
  `lookup_species_count()`. The resulting output is saved as a .csv file to
  a directory specified by the user, and can be passed as an argument to
  to `build_email()`
}
