% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_occurrences.R
\name{filter_occurrences}
\alias{filter_occurrences}
\title{Filter species occurrence data by state, LGA and/or shapefile}
\usage{
filter_occurrences(species_records, shapes_path = NULL)
}
\arguments{
\item{species_records}{A data.frame or tibble as produced by
`search_occurrences()` or otherwise. Requires at a minimum numerical
columns `decimalLongitude`, `decimalLatitude` , and character columns
`cl10923`, `state`, `lga`, and `shape`.}

\item{shapes_path}{A character string specifying the directory that includes
folders containing each shapefile specified in the `shape` column of
`occ_list`. Each shapefile belongs in its own folder, and all files in
that folder must be named identically to the folder name except for the
file suffixes. Each shapefile must contain a feature named `SHAPE_NAME`.
Defaults to `NULL` i.e. no shapefiles provided.}
}
\value{
A tibble containing provided data for each occurrence record,
   filtered to only include rows for those with state, LGA or shapefile
   matches. Contains __ ALA-specific columns with data regarding taxonomy,
   location and record data, 7 user-supplied columns from `species_list`,
   logical columns for each list as per `species_list`, and a column
   indicating the Australian state/territory of the location.
}
\description{
This function performs the location filtering of species records downloaded
   by `search_occurrences()` or otherwise. Depending on the provided location
   columns, filtering will be performed by `state`, `LGA` and/or `shapefile`.
   Only occurrences that are located within the state, LGA or shapefile
   boundaries are retained in the output tibble,
}
\details{
Currently, the provided tibble should ideally be produced by
   `search_occurrences()`, which itself modifies output from
   `galah::atlas_occurrences()`. Location filtering requires the presence of
   {galah} fields `decimalLongitude`, `decimalLatitude`, `cl10923` (2018 LGA
   boundaries) and provided list fields `shape`, `lga` and `shape`. Location
   matches are only identified with verbatim matches of state abbreviations,
   LGA names and shapefile names.

The Australian state/territory of each occurrence is currently identified
   using the Australian Coastal Waters Act 1980, which defines both
   terrestrial and maritime jurisdictions of rhe Australian states and
   territories.
}
