% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_species_lists.R
\name{get_species_lists}
\alias{get_species_lists}
\title{Import all files, formatting with logical columns stating source}
\usage{
get_species_lists(df)
}
\arguments{
\item{df}{A `data.frame` containing columns 'path' and 'label'}
}
\description{
Early days
}
