% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_species_lists2.R
\name{get_species_lists2}
\alias{get_species_lists2}
\title{Import all files, formatting with logical columns stating source}
\usage{
get_species_lists2(lists_df)
}
\arguments{
\item{df}{A `data.frame` containing columns 'path' and 'label'}
}
\description{
Redoing Martin's function minus lapply and to hopefully run smoother
}
