% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_occurrences.R
\name{identify_aus}
\alias{identify_aus}
\title{Identify whether each record in a dataframe sits in Australian territory}
\usage{
identify_aus(species_records)
}
\arguments{
\item{species_records}{A dataframe/tibble prodcued by `atlas_occurrences()` or
otherwise, containing at minimum columns containing longitude and latitude
columns. Each row represents a unique occurrence.}
}
\value{
Returns the exact provided dataframe with non-Australian occurrences
   removed.
}
\description{
When provided with some (potentially modified) dataframe/tibble as produced
   by `atlas_occurrences()` or other, occurrence rows will be filtered out if
   they do not sit in Australian terrestrial or maritime territory as defined
   by the Seas and Submerged Lands Act (SSLA) 1973 and subsequent treaties.
   This relies on the presence of numeric columns `decimalLongitude` and
   `decimalLatitude` (default {galah} coordinate columns) to match the
   coordinates of the occurrences to Australian territory boundaries.
}
