% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_occurrences.R
\name{identify_lga}
\alias{identify_lga}
\title{Identify matches for the Australian LGA of each record in a dataframe}
\usage{
identify_lga(species_records)
}
\arguments{
\item{species_records}{A dataframe/tibble prodcued by `atlas_occurrences()` or
otherwise, containing at minimum columns containing longitude and latitude
columns. Each row represents a unique occurrence.}
}
\value{
Returns the exact provided dataframe with an additional logical
   column `flagged_lga` indicating a match to the .
}
\description{
When provided with some (potentially modified) dataframe/tibble as produced
   by `atlas_occurrences()` or other, a new logical column `flagged_lga` will
   be created and appended to the end of the dataframe, identifying whether
   the Australian LGA each occurrence sits in (as provided by the ALA field
   `cl10923`) matches the provided LGA in the `lga` column.
}
