% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_occurrences.R
\name{identify_shape}
\alias{identify_shape}
\title{Identify the presence of species occurrences in a set of shapefiles}
\usage{
identify_shape(species_records, shapes_path = NULL)
}
\arguments{
\item{species_records}{A dataframe/tibble prodcued by `atlas_occurrences()` or
otherwise, containing at minimum columns containing longitude and latitude
columns, and a column named `shape` specifying the name of the shapefiles
to filter each species. Multiple shapefiles may be used, but it is limited
to one shapefile per row. Each row represents a unique occurrence.}

\item{shapes_path}{A character string specifying the directory that includes
folders containing each shapefile specified in the `shape` column of
`species_records`. Each shapefile belongs in its own folder, and all files in
that folder must be named identically to the folder name except for the
file suffixes. Each shapefile must contain a feature named `SHAPE_NAME`.}
}
\value{
Returns the exact provided dataframe with an additional character
   column `shape_feature` containing the name of the provided shapefile
   feature if the occurrence does sit within the region specified by the
   shapefile.
}
\description{
Within the {koel} `get_occurrences()` workflow, species may be provided with
   an optional `shape` argument specifying the name of the shapefile inside
   which occurrences of that species should be kept. When provided with some
   (potentially modified) dataframe/tibble as produced by
   `atlas_occurrences()` or other, a new column `shape_feature` will be
   created and appended to the end of the dataframe, identifying the feature
   of its specified shapefile that each occurrence occurs in, if it does
   occur in the shapefile. If an occurrence does not occur in the specified
   shapefile then `NA` is returned.
}
