% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_occurrences.R
\name{identify_state}
\alias{identify_state}
\title{Identify the Australian state of each record in a dataframe}
\usage{
identify_state(occ_list)
}
\arguments{
\item{occ_list}{A dataframe/tibble prodcued by `atlas_occurrences()` or
otherwise, containing at minimum columns containing longitude and latitude
columns. Each row represents a unique occurrence}
}
\value{
Returns the exact provided dataframe with an additional character
   column `cw_state` (Coastal Waters state) containing Australian state
   abbreviations
}
\description{
When provided with some (potentially modified) dataframe/tibble as produced
   by `atlas_occurrences()` or other, a new column `cw_state` will be created
   and appended to the end of the dataframe, identifying the Australian state
   each occurrence sits in. This relies on the presence of numeric columns
   `decimalLongitude` and `decimalLatitude` (default {galah} coordinate
   columns) to match the coordinates of the occurrences to the Australian
   state boundaries as described and delineated by the Coastal Waters Act
   1980. If an occurrence does not occur in any state jurisdiction then
   `NA` is returned.
}
