% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koel_defensive.R
\name{koel_defensive}
\alias{koel_defensive}
\title{Internal function to perform defensive programming checks}
\usage{
koel_defensive(...)
}
\description{
This function is to be called only at the beginning of koel functions. It
   performs defensive programming through a set of if/else statements that
   provide informative errors. The main purpose of this function is to
   contain and aggregate all defensive programming in the one location.
}
\details{
Just need to provide the following code at the beginning of each function:

```
this_call <- match.call(expand.dots = TRUE)
this_call[[1]] <- as.name("koel_defensive")
eval.parent(this_call)
```
}
