% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ala_records.R
\name{lookup_species_count}
\alias{lookup_species_count}
\title{Look up and append counts of species records}
\usage{
lookup_species_count(species_list, filter_df, max_counts)
}
\arguments{
\item{species_list}{A data.frame or tibble containing information on species
of biosecurity interest. Must contain character columns 'correct_name',
'search_term', and 'common_name' for each species, and logical columns for
each alerts list indicating whether or not the species appears on a list.
May be produced by `get_species_list2()`.}

\item{filter_df}{A tibble specifying {galah} query parameters, as produced
by `galah::galah_filter()`. May be produced by `build_galah_query()`.}

\item{max_counts}{A numeric upper bound. Species with record counts greater
then this value will be excluded from the final output.}
}
\value{
A tibble based on the input `species_list`, with an additional column
  named `counts`
}
\description{
This function searches the ALA for records of species of biosecurity
  interest within a specified time period using the {galah} package. The
  output of this function may be passed as an argument to `download_records()`.
}
