% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_occurrences.R
\name{search_name_fields}
\alias{search_name_fields}
\title{Search ALA with multiple search terms and fields}
\usage{
search_name_fields(field, start_date, end_date, search_terms)
}
\arguments{
\item{field}{Single character designating field to be searched in galah e.g.
`"scientificName"`, `"genus"`}

\item{start_date}{Date to begin search of ALA occurrences. `character` object
and should be in the form `"YYYY-mm-ddTHH::MM::SSZ"`.}

\item{end_date}{Date to end search of ALA occurrences. `character` object
and should be in the form `"YYYY-mm-ddTHH::MM::SSZ"`.}

\item{search_terms}{Character vector (may be of length 1) of search terms to
be provided to `galah_filter()`. Length should be capped at 100 to avoid
`{galah}` errors.}
}
\description{
Helper function to perform galah searches for given search fields
}
