% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_occurrences.R
\name{search_occurrences}
\alias{search_occurrences}
\title{Search and download occurrence data for a species list}
\usage{
search_occurrences(species_list, common_names, start_date, end_date = 0)
}
\arguments{
\item{species_list}{A data.frame or tibble containing information on species
of biosecurity interest. Must contain character columns `correct_name`,
`search_term`, and `common_name` for each species, and logical columns for
each alerts list indicating whether or not the species appears on a list.
May be produced by `get_species_list2()`.}

\item{common_names}{A data.frame or tibble containing two character columns,
`correct_name` and `common_name`, and rows indicating the accepted common
name for each correct species name. May be produced by
`common_names_assigned()`.}

\item{start_date}{Date to begin search of ALA occurrences. May be in one of
two forms: either a single dbl indicating how many days prior from the current
date to begin the search, or a character vector indicating the date to
search from in `"ddmmyyyy"` format.}

\item{end_date}{Date to end search of ALA occurrences. May be in one of
two forms: either a single dbl indicating how many days prior from the current
date to end the search, or a character vector indicating the date to
search to in `"ddmmyyyy"` format. Default value is 0 i.e. search up to
the current date.}
}
\value{
A tibble containing the downloaded data for each occurrence record.
   Contains 32 ALA-specific columns with data regarding taxonomy, location,
   media, uploading user, data type; 5 user-supplied columns from `species_list`
   containing correct, provided, common and searched names and jurisdictions,
   logical columns for each list as per `species_list`, and a column
   indicating the Australian state/territory jurisdiction each occurrence was
   located as per the Australian Coastal Waters Act 1980.
}
\description{
This function performs the search processes for the identification of species
   records from a search-term dataframe. It searches within some specified
   time period for exact name matches using {galah}, and outputs occurrence,
   taxonomic and location data for each occurrence within the timeframe.
}
\details{
The searches are performed for each block of 100 search terms, and
   relevant text output (in between `"Checking queue"` outputs) indicate the
   number of records downloaded for each block of 100 terms, followed by the
   number of media files downloaded.
}
