% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_email.R, R/send_email.R
\name{send_email}
\alias{send_email}
\title{Function to send html tables of occurrences in emails to stakeholders}
\usage{
send_email(recipients, output_file, subject = "ALA Biosecurity Alert")

send_email(recipients, output_file, subject = "ALA Biosecurity Alert")
}
\arguments{
\item{recipients}{}

\item{output_file}{A `character string` providing the path to the outputted
html file containing the {gt} table to be rendered in the email. This
path is produced in `build_email()` but can be provided separately too.}

\item{subject}{An optional `character string` of the subject of the email.
If not provided, defaults to "ALA Biosecurity Alerts".}
}
\value{
No object is returned. This function exists only to send an email
   containing the relevant tables for a biosecurity alert.
}
\description{
This function is a wrapper for a set of {emayili} functions that are used to
   send biosecurity alert emails to addressees of each provided list. It
   renders a previously created and saved html file produced with RMarkdown
   of occurrences of species of interest. Is used internally in `build_email()`
   but can be deployed externally if necessary.
}
