% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALA_record_download.R
\name{ala_record_download}
\alias{ala_record_download}
\title{Download ALA records for species identified with `ala_species_counts`}
\usage{
ala_record_download(species_list, common_names, filter_df, path)
}
\arguments{
\item{species_list}{A `data.frame` or `tibble` preferably produced by
`ala_species_counts()`, with each row representing a search term for a species
with records in the ALA in some set timeframe. Must contain character columns
'correct_name', 'search_term' and 'common_name', logical columns indicating
the presence of that species on each alerts list, and a numerical column
indicating the number of records for each species.}

\item{common_names}{A `data.frame` or `tibble` preferably produced by
`common_names_assigned()`, each row representing the accepted common_name for
each correct species name. Should contain only two character columns,
'correct_name' and 'common_name'.}

\item{filter_df}{A data.frame or tibble of ALA query conditions, as produced
by `galah::galah_filter()`. May be produced by `build_ala_query()`.}

\item{path}{A character string specifying the (temporary) cache directory to
save downloaded media files for each occurrence, and the downloaded occurrence
data. The string must end in "/". The path must describe an existing directory,
and if no 'species_images' folder exists within this directory then one will
be created, in which the media output will be saved.}
}
\value{
A tibble containing the downloaded data for each occurrence record.
   Contains 30 ALA specific columns with data regarding location, media,
   uploading user and data type, 2 columns containing correct and common names,
   and logical columns for each list indicating the presence of that species
   on that list.
}
\description{
This function downloads data of individual species records identified by
   `ala_species_counts()` and the media associated with these records. The
   function returns a tibble of this data, with a variety of variables related
   to the location, media and taxonomy of the observation. This tibble is
   saved as a .csv file to a chosen directory, and can be passed as an argument
   to `build_email()`
}
