% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALA_species_counts.R
\name{ala_species_counts}
\alias{ala_species_counts}
\title{Search all of ALA for species and record their counts of records in the timeframe}
\usage{
ala_species_counts(species_list, filter_df, max_counts)
}
\arguments{
\item{species_list}{A `data.frame` containing columns of 'correct_name', 'search_term', 'common_name' for species, and T/F columns for each list}

\item{filter_df}{A `list` of ALA query conditions}

\item{max_counts}{A `dbl` indicating the maximum number of counts for a species that we still want a report for}
}
\description{
Redoing Martin's function minus lapply and to hopefully run smoother
}
