% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_species_lists.R
\name{get_species_lists}
\alias{get_species_lists}
\title{Import user-supplied species lists from a single directory}
\usage{
get_species_lists(lists_path, list_suffix = "_list", synonym_delimiter = ",")
}
\arguments{
\item{lists_path}{Path to the directory where lists are saved, ending in `"/"`.}

\item{list_suffix}{Character string (case insensitive) between the list name
and file extension.}

\item{synonym_delimiter}{An optional character string detailing the delimiter
used for multiple synonyms in the synonym column of the lists. Defaults to
`", "` or `","`.}
}
\value{
A data.frame of unique scientific names, the search term used to
  match those names to the ALA taxonomy, a common name for each species, the
  state and LGA jurisdictions of interest for each species, and a column for
  the name of the list that name was provided by. This data.frame may be
  passed to `assign_common_names()` and `search_occurrences()`.
}
\description{
This function takes a small data.frame produced by `collate_lists()` and
   facilitates the import of species list files described there. The rest of
   the function utilises {dplyr} and {tidyr} functions to clean and combine
   the species lists into a single data.frame suitable for use with {galah}.
}
