% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ala_records.R
\name{lookup_species_count}
\alias{lookup_species_count}
\title{Look up and append counts of species records}
\usage{
lookup_species_count(species_list, max_counts, start_date, end_date = 0)
}
\arguments{
\item{species_list}{A data.frame or tibble containing information on species
of biosecurity interest. Must contain character columns 'correct_name',
'search_term', and 'common_name' for each species, and logical columns for
each alerts list indicating whether or not the species appears on a list.
May be produced by `get_species_list2()`.}

\item{max_counts}{A numeric upper bound. Species with record counts greater
then this value will be excluded from the final output.}

\item{start_date}{Date to begin search of ALA occurrences. May be in one of
two forms: either a `dbl` indicating how many days prior from the current
date to begin the search, or a `character` vector indicating the date to
search from in "ddmmyyyy" format.}

\item{end_date}{Date to end search of ALA occurrences. May be in one of
two forms: either a `dbl` indicating how many days prior from the current
date to end the search, or a `character` vector indicating the date to
search to in "ddmmyyyy" format. Default value is 0 i.e. search up to
the current date}
}
\value{
A tibble based on the input `species_list`, with an additional column
  named `counts`
}
\description{
This function searches the ALA for records of species of biosecurity
  interest within a specified time period using the {galah} package. The
  output of this function may be passed as an argument to `download_records()`.
}
