% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_occurrences.R
\name{search_occurrences}
\alias{search_occurrences}
\title{Search and download occurrence data for a species list}
\usage{
search_occurrences(
  species_list,
  common_names,
  event_date_start,
  event_date_end = 0,
  upload_date_start = "01-01-1570",
  upload_date_end = 0
)
}
\arguments{
\item{species_list}{A data.frame or tibble containing information on species
of biosecurity interest. Must contain character columns `correct_name`,
`search_term`, and `common_name` for each species, and logical columns for
each alerts list indicating whether or not the species appears on a list.
May be produced by `get_species_list2()`.}

\item{common_names}{A data.frame or tibble containing two character columns,
`correct_name` and `common_name`, and rows indicating the accepted common
name for each correct species name. May be produced by
`common_names_assigned()`.}

\item{event_date_start, event_date_end}{Dates to begin and end ALA occurrences
filter for event date field. Allows filtering of records that have been
uploaded in some time period. May be in one of two forms: either a single
dbl indicating how many days prior from the currentdate to begin the
search, or a character vector indicating the date to search from in
`"ddmmyyyy"` format.  `event_date_end` defaults to 0 (current date).}

\item{upload_date_start, upload_date_end}{Dates to begin and end ALA
occurrences filter for upload date field. Allows filtering of records that
have been uploaded in some time period. May be in one of two forms: either
a single dbl indicating how many days prior from the current date to begin
the search, or a character vector indicating the date to search from in
`"ddmmyyyy"` format. `upload_date_start` defaults to `"01-01-1570"` and
`upload_date_end` defaults to 0 (i.e. current date).}
}
\value{
A tibble containing the downloaded data for each occurrence record.
   Contains 32 ALA-specific columns with data regarding taxonomy, location,
   media, uploading user, data type; 5 user-supplied columns from `species_list`
   containing correct, provided, common and searched names and jurisdictions,
   logical columns for each list as per `species_list`, and a column
   indicating the Australian state/territory jurisdiction each occurrence was
   located as per the Australian Coastal Waters Act 1980.
}
\description{
This function performs the search processes for the identification of species
   records from a search-term dataframe. It searches within some specified
   time period for exact name matches using {galah}, and outputs occurrence,
   taxonomic and location data for each occurrence within the timeframe.
}
\details{
The searches are performed for each block of 100 search terms, and
   relevant text output (in between `"Checking queue"` outputs) indicate the
   number of records downloaded for each block of 100 terms, followed by the
   number of media files downloaded.
}
